api/requirements.txt
opencv-python
numpy
Pillow
matplotlib
tqdm
torch
scikit-image
torchvision
monai
torchmetrics
scikit-learn
pytorch_lightning
imagecodecs
fastapi
python-multipart
uvicorn